+++ backend/api/gamaharului/src/main/java/com/tibelian/gamaharului/repository/TrackRepository.java
@@ -2,9 +2,9 @@ package com.tibelian.gamaharului.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.tibelian.gamaharului.model.Track;
+import com.tibelian.gamaharului.model.music.Track;
 
-public interface TrackRepository extends JpaRepository<Track, Long> {
+public interface TrackRepository extends JpaRepository<Track, Integer> {
